[FEATURE] Add M3 S0 to beginning of file
Summary:
Looking through my test files again, I saw that we had decided on the
forum post that we always need an `M3 S0` line at the beginning of the
file.

Instead of trying to convey to everyone that they need to edit their
start gcode in lightburn, I'm just going to prepend it on the file when
fixing it so that it works out of the box.

Test plan:
 - Ran the right-click setup in windows
 - Ran the script using WSL in windows

Both produced a file with `M3 S0` at the top of the file